removed globals and created structure
